refactor: rename artifact_management to artifact_service
- Renamed the 'artifact_management' directory to 'artifact_service' to better reflect its role as the infrastructure layer.
- Renamed 'service.py' to 'gcs_service.py' to explicitly define the underlying technology.
- Dropped schemas.py from the service layer (will be decoupled into the tools layer).
- Updated 'app_builder.py' to wire up the new gcs_service path.
